requirements.txt
# FastAPI and Web
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
pymongo>=4.4.1
mongoengine>=0.27.0
dnspython>=2.3.0

# AI and ML
google-generativeai>=0.1.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
mongomock>=4.1.2

# Utils
python-dotenv>=1.0.0
pydantic>=2.1.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
pytz>=2023.3
